+++ src/main/java/ca/gc/aafc/collection/api/entities/CollectingEvent.java
@@ -1,8 +1,8 @@
 package ca.gc.aafc.collection.api.entities;
 
-import ca.gc.aafc.collection.api.datetime.ISODateTime;
 import ca.gc.aafc.collection.api.dto.GeoreferenceAssertionDto;
 import ca.gc.aafc.dina.entity.DinaEntity;
+import ca.gc.aafc.dina.datetime.ISODateTime;
 
 import com.vladmihalcea.hibernate.type.array.ListArrayType;
 import com.vladmihalcea.hibernate.type.array.StringArrayType;
